refactor: add UserDataFactory class instead of helper
